Create Learning Style Quiz.mlapp
This was a culmination of our group work from Friday onwards. At our first work session meet up, we all spent time researching and looking up information and playing around on our individual MATLAB's to see if we could get a framework of our quiz to function. I got a basic outline to work (getting the code to respond to buttons being pressed and moving from one multiple choice question to another). The results page is not working still, so Nash and Shreya are going to work on getting that to function, editing and cleaning up my code, and continuing the code.