Working Results Panel
I spent a few hours researching and experimenting to try to get a working results function. Initially, it was just popping up as a message and ending the quiz, but we wanted a more interactive page that was detail-heavy for our results. I figured out to toggle the visibility of certain elements, so the results panel can now show up at the end of the quiz. The code is still very basic and needs to be edited/cleaned up and fleshed out as a group.