Slider Question Type
This is the culmination of 4 days worth of work in trying to add a new question type for the quiz. I have been working on adding a slider type of question where a prompt is given and participants get to rank their affinity towards the prompt. Though the basics of it are working, there are still some errors to be sorted out regarding when the slider appears, which we will continue to work on.