Editing Audio Code
After our group meeting, I worked on editing my code for the audio to adapt to Shreya's code for the slider question. I also added a play button so that the user can play the audio as many times as necessary to answer the question, as originally it would only play once.